af_ktls: Add retransmit_hint to speed up retranmissions
Keep a retransmit_hint record in the TLS context.
When we do a linear search to find a record for resync we move
this hint pointer forward.
This Save us traversing the record list many times if we have
many retranmissions and we jump back and forward in the retranmission
queue.

Issue: None
Change-Id: Ic0a8cf4c32d6c0688bf5bd27a10ca4fbaf233ed7
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>